promoted glacier_melt_ts to time-series
